+++ dbbackup/utils.py
@@ -278,6 +278,7 @@ def uncompress_file(inputfile, filename):
     """
     zipfile = gzip.GzipFile(fileobj=inputfile, mode="rb")
     try:
+        inputfile.seek(0)
         outputfile = create_spooled_temporary_file(fileobj=zipfile)
     finally:
         zipfile.close()