+++ docs/changelog.rst
@@ -7,6 +7,7 @@ Unreleased
 * Add PostgreSQL Schema support by @angryfoxx in https://github.com/jazzband/django-dbbackup/pull/507
 * Fix restore of database from S3 storage by reintroducing inputfile.seek(0) to utils.uncompress_file
 * Fix bug where dbbackup management command would not respect settings.py:DBBACKUP_DATABASES
+* Add option `--if-exists` for pg_dump command
 
 4.1.0 (2024-01-14)
 ------------------